added optionHandler txs
