+++ src/OptionHandler.ts
@@ -1,1 +1,15 @@
-export { handleOptionsBought, handleOptionsSold } from "./OptionExchange";
+
+import { addOptionsBoughtAction, addOptionsSoldAction } from "./helper";
+
+import {
+  OptionsBought,
+  OptionsSold
+} from "../generated/OptionExchange/OptionExchange";
+
+export function handleOptionsBought(event: OptionsBought): void {
+  addOptionsBoughtAction(event)
+}
+
+export function handleOptionsSold(event: OptionsSold): void {
+  addOptionsSoldAction(event)
+}
