+++ src/helper.ts
@@ -1,11 +1,14 @@
-import { BigDecimal, BigInt, Address } from "@graphprotocol/graph-ts";
+import { BigDecimal, BigInt, Address, ethereum } from "@graphprotocol/graph-ts";
 
 import {
   Account,
   LongPosition,
   ShortPosition,
-  Stat
+  Stat,
+  OptionsBoughtAction,
+  OptionsSoldAction
 } from "../generated/schema";
+import { OptionsBought, OptionsSold } from "../generated/OptionExchange/OptionExchange";
 
 export let BIGINT_ONE = BigInt.fromI32(1);
 export let BIGINT_ZERO = BigInt.fromI32(0);
@@ -382,6 +385,110 @@ export function updateLiquidatedPosition(
   liquidateActions.push(liquidationId);
   position.liquidateActions = liquidateActions;
   position.save();
+}
+
+export function addOptionsBoughtAction(
+  event: OptionsBought,
+): void {
+
+  const txHash = event.transaction.hash.toHex();
+
+  const otoken = event.params.series.toHex();
+  const buyer = event.params.buyer;
+  const amount = event.params.optionAmount;
+  const premium = event.params.premium;
+  const fee = event.params.fee;
+  const timestamp = event.block.timestamp;
+  const from = event.transaction.from;
+  const receipt = event.receipt;
+  const txLogs = receipt ? receipt.logs : [];
+
+  const id = txHash + "-" + otoken;
+
+  const optionsBoughtAction = new OptionsBoughtAction(id);
+
+  optionsBoughtAction.otoken = otoken;
+  optionsBoughtAction.buyer = buyer;
+  optionsBoughtAction.amount = amount;
+  optionsBoughtAction.premium = premium;
+  optionsBoughtAction.fee = fee;
+  optionsBoughtAction.timestamp = timestamp;
+  optionsBoughtAction.transactionHash = txHash;
+
+  optionsBoughtAction.save();
+
+  const total = premium.plus(fee);
+
+  // add fees to stats
+  updateStats(amount, fee, true);
+
+  for (let i = 0; i < txLogs.length; ++i) {
+    // if event is to Controller, avoid reading all events
+    if (txLogs[i].address.toHexString() == CONTROLLER) {
+      // if topic is ShortOtokenBurned and account owner is tx sender (trader)
+      if (
+        txLogs[i].topics[0].toHexString() == SHORT_OTOKEN_BURNED &&
+        txLogs[i].topics[2].toHexString().slice(26) ==
+          from.toHexString().slice(2)
+      ) {
+        // if topic is shortOTokenBurned and account owner is tx sender (trader)
+        updateOptionShortPosition(true, buyer, otoken, amount, id, total);
+        return;
+      }
+    }
+  }
+  updateOptionLongPosition(true, buyer, otoken, amount, id, total);
+}
+
+export function addOptionsSoldAction(
+  event: OptionsSold,
+): void {
+
+  const txHash = event.transaction.hash.toHex();
+
+  const receipt = event.receipt;
+  const txLogs = receipt ? receipt.logs : [];
+
+  const seller = event.params.seller;
+  const otoken = event.params.series.toHex();
+  const amount = event.params.optionAmount;
+
+  const id = txHash + "-" + otoken;
+
+  const optionsSoldAction = new OptionsSoldAction(id);
+
+  optionsSoldAction.otoken = otoken;
+  optionsSoldAction.seller = seller;
+  optionsSoldAction.amount = amount;
+  optionsSoldAction.premium = event.params.premium;
+  optionsSoldAction.fee = event.params.fee;
+  optionsSoldAction.timestamp = event.block.timestamp;
+  optionsSoldAction.transactionHash = event.transaction.hash.toHex();
+
+  optionsSoldAction.save();
+
+  const total = event.params.premium.minus(event.params.fee);
+
+  // add fees to stats
+  updateStats(event.params.optionAmount, event.params.fee, false);
+
+  for (let i = 0; i < txLogs.length; ++i) {
+    // if event is to Controller, avoid reading all events
+    if (txLogs[i].address.toHexString() == CONTROLLER) {
+      // if topic is ShortOtokenMinted and account owner is tx sender (trader)
+      if (
+        txLogs[i].topics[0].toHexString() == SHORT_OTOKEN_MINTED &&
+        txLogs[i].topics[2].toHexString().slice(26) ==
+          event.transaction.from.toHexString().slice(2)
+      ) {
+        // if topic is shortOtokenMinted and account owner is tx sender (trader)
+        updateOptionShortPosition(false, seller, otoken, amount, id, total);
+        return;
+      }
+    }
+  }
+  updateOptionLongPosition(false, seller, otoken, amount, id, total);
+  
 }
 
 // dashboard functions
